Move operation unit tests to operations.
